requirements.txt
fastapi
sqlmodel
uvicorn
dotenv
pyyaml
pytz
pytest
httpx
psycopg2-binary
